Change a level compilation system
To maximize a performance I decided that rendering every single cube separately takes to much resourses from the system, so I made a compilation script. It takes all coordinates of a level, splits them in several planes and makes several PNG images of these planes. Then it writes it's relative coordinates in a JSON file, that will be passed to a game when loaded
Script is a little broken at the moment, but I hope to fix it in a near future.